requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
langchain
langchain-openai
langchain-community
langchain-experimental
qdrant-client
python-dotenv
jinja2
pydantic-settings
python-multipart